Revert "[ObjC]fix 64 to 32 bit clang conversion warning in src/core/lib (#29877)" (#31843)
This reverts commit eb0a5918116fa7c600ab1e2fb47782a5c55c55b9.

+++ src/core/ext/transport/chttp2/transport/flow_control.cc
@@ -294,7 +294,7 @@ void TransportFlowControl::UpdateSetting(
         urgency = FlowControlAction::Urgency::UPDATE_IMMEDIATELY;
       }
       *desired_value = new_desired_value;
-      (action->*set)(urgency, static_cast<uint32_t>(*desired_value));
+      (action->*set)(urgency, *desired_value);
     }
   } else {
     int64_t delta = new_desired_value - *desired_value;
